feat: add scripts for vendor asset management and database maintenance
- Introduced `audit_bundle_budget.mjs` to enforce bundle size limits for JavaScript assets.
- Added `audit_secrets.py` to scan for potential secrets in source files.
- Created `audit_vendor_assets.mjs` to validate vendored assets against a manifest.
- Implemented `backup_database.py` for creating online backups of the SQLite database.
- Developed `check_database.py` for running integrity checks on the SQLite database.
- Added `generate_vendor_sbom.mjs` to generate a Software Bill of Materials (SBOM) for vendored assets.
- Created `restore_database.py` to verify and restore backups of the SQLite database.
- Added `vendor-manifest.json` to track vendored assets, including their licenses and sources.
- Introduced `vendor.cdx.json` for CycloneDX SBOM format to document vendored components.

+++ backend/documents/docx_bid_context_service.py
@@ -313,7 +313,7 @@ def enrich_context_with_lot_summaries(context):
                 'ds_phan_lo': []
             })
             try:
-                group['_tong_gia_tri_trung_thau_raw'] += float(winner_item.get('_gia_trung_thau_raw') or 0)
+                group['_tong_gia_tri_trung_thau_raw'] += int(winner_item.get('_gia_trung_thau_raw') or 0)
             except Exception:
                 pass
             won_lot_item = {
@@ -344,7 +344,6 @@ def enrich_context_with_lot_summaries(context):
     context['so_phan_lo_khong_co_nha_thau_tham_du'] = len(lots_without_participants)
     context['so_phan_lo_co_nha_thau_trung'] = len(lots_with_winner)
     context['so_phan_lo_tham_du_khong_trung'] = len(lots_participated_without_winner)
-
